requirements.txt
fastapi>=0.100.0
uvicorn>=0.34.0
python-dotenv>=1.1.0
python-multipart>=0.0.20
motor>=3.7.0
aiohttp>=3.11.0
pytz>=2024.1
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
tensorflow>=2.15.0
torch>=2.2.0
transformers==4.9.2
pymongo==3.12.0
python-multipart==0.0.5
passlib==1.7.4
python-jose==3.3.0
beautifulsoup4==4.9.3
aiohttp==3.7.4
Pillow==8.3.2